% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atus-data.R
\docType{data}
\name{atusresp}
\alias{atusresp}
\title{ATUS 2014 Respondents}
\format{A data frame with 11,592 observations on the following variables:
\describe{
 \item{tucaseid}{respondent id}
 \item{earn_week_overtime}{total weekly overtime earnings}
 \item{hourly_wage}{hourly rate of pay on your main job excluding overtime pay, tips, and commissions}
 \item{hrs_weekly}{hours usually worked per week at this rate}
 \item{hourly_status}{paid hourly or nonhourly}
 \item{weekly_earn}{weekly earnings at main job}
 \item{other_payment}{whether respondent typically receives overtime pay, commissions, or tips at the main job}
 \item{weeks_year}{weeks paid per year}
 \item{hrs_weekly2}{hours worked per week at main job}
 \item{hrs_weekly_other}{hours worked per week at other jobs}
 \item{industry_code}{industry code (see codebook)}
 \item{occup_code}{occupation code (see codebook)}
 \item{labor_status}{labor force status}
 \item{student_status}{whether or not the respondent is currently enrolled in a high school, college, or university}
 \item{num_children}{number of children in household}
 \item{holiday}{whether the diary day was a holiday}
 \item{diary_date}{date of diary day in 'yyyymmdd' format}
 \item{diary_day}{day of week for diary day}
 \item{wt}{ATUS final weight}
}}
\source{
\url{http://www.bls.gov/tus/datafiles_2014.htm} with code book at \url{http://www.bls.gov/tus/atusintcodebk14.pdf}.
}
\usage{
atusresp
}
\description{
This is a 11,592 by 19 data frame containing information on the respondents, including their labor force status, earnings, and the diary day (day on which they were interviewed).
}
\examples{
data(atusresp)

}
\keyword{datasets}

