% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atus-data.R
\docType{data}
\name{atusact}
\alias{atusact}
\title{ATUS 2014 Activities}
\format{A data frame with 223,249 observations on the following variables:
\describe{
 \item{tucaseid}{respondent ID}
 \item{actid}{activity number}
 \item{location}{where the activity was performed (see codebook)}
 \item{dur}{Duration of activity in minutes (last activity truncated at 4:00 a.m.)}
\item{cum_dur}{Cumulative duration of activity lengths in minutes; last activity truncated at
4:00am or 1440 minutes; cumulative total of `duration` for each individual/activity combination}
\item{starttime}{Activity start time.}
\item{stoptime}{Activity stop time.}
\item{t1code}{Tier 1 activity code (see codebook for details).}
\item{t2code}{Tier 2 activity code (see codebook for details).}
\item{t3code}{Tier 3 activity code (see codebook for details).}

}}
\source{
\url{http://www.bls.gov/tus/datafiles_2014.htm} with code book at \url{http://www.bls.gov/tus/atusintcodebk14.pdf}.
}
\usage{
atusact
}
\description{
This is a 223,249 by 10 data frame relating to how ATUS respondents spent their diary day. It includes information such as activity codes, activity start and stop times, and locations. Due to the presence of multiple codes for things like activities and locations, the codes have not been converted to their appropriate labels for conciseness. This file differs from the original Activities file in that all variables relating to dependent care have been removed. Durations for events with a reported end time after 4am the following day have been truncated so that the cumulative duration of all activities do not exceed 24 hours for any individual.
}
\examples{
data(atusact)

}
\keyword{datasets}

