% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atus-data.R
\docType{data}
\name{atuscps}
\alias{atuscps}
\title{CPS 2014}
\format{Contains information about each household member of all individuals selected to participate in ATUS.
A data frame with 68,198 observations on the following variables:
\describe{
 \item{tucaseid}{respondent id}
 \item{lineid}{household member id}
 \item{region}{region of household (definitions are listed in Appendix A of codebook)}
 \item{division}{census division of  household (definitions are listed in Appendix A of codebook)}
 \item{famincome}{family income}
 \item{housing_type}{type of housing unit}
 \item{hh_size}{number of members in household}
 \item{month}{month of interview}
 \item{edu}{respondent education level, condensed}
 \item{sex}{respondent sex}
 \item{age}{respondent age}
 \item{citizen}{citizenship status of respondent}
 \item{atus}{indicator for whether respondent was interviewd for ATUS}
}}
\source{
\url{http://www.bls.gov/tus/datafiles_2014.htm}
}
\usage{
atuscps
}
\description{
This is a 68,198 by 13 data frame containing information about each household member of all individuals in ATUS. The information here was merged in from the CPS and was collected 2 to 5 months before the ATUS interview.
}
\examples{
data(atuscps)

}
\keyword{datasets}

