% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atus-data.R
\docType{data}
\name{atussum}
\alias{atussum}
\title{ATUS 2014 Activity Summary}
\format{A data frame with 11,592 observations on the following variables:
\describe{
 \item{tucaseid}{respondent id}
 \item{wt}{final ATUS weight}
 \item{age}{respondent age}
 \item{sex}{respondent sex}
 \item{edu}{respondent education level, condensed}
 \item{race}{respondent race, condensed}
 \item{hispanic}{whether the respondent identifies as Hispanic}
 \item{metro}{metropolitan status}
}
These variables are then followed by 384 variables relating to how much time each individual spent total on each of the activities, where each variable corresponds to a specific activity, denoted by its tier 1, tier 2, and tier 3 codes concatenated in its variable name.}
\source{
\url{http://www.bls.gov/tus/datafiles_2014.htm} with code book at \url{http://www.bls.gov/tus/atusintcodebk14.pdf} and \url{http://www.bls.gov/tus/atuscpscodebk14.pdf}.
}
\usage{
atussum
}
\description{
This is a 11,592 by 393 data frame containing information on te total time each ATUS respondent spent doing each activity on the diary day.
}
\examples{
data(atussum)

}
\keyword{datasets}

